Predict depth from a ratio raster or table

Applies a fitted fit_bathymetry() model to new ratio values to obtain predicted depths. When newdata is a raster the result is a depth map.

Usage

predict_bathymetry(model, ratio)

Arguments

model

A satbathy_model from fit_bathymetry().

ratio

A single-layer terra::SpatRaster of ratio values (its layer is treated as the ratio predictor).

Value

A single-layer terra::SpatRaster named "depth".

See also

Examples

img <- terra::rast(system.file("extdata", "example_sr.tif",
                               package = "satbathy"))
pts <- terra::vect(system.file("extdata", "example_points.gpkg",
                               package = "satbathy"))
img <- terra::mask(img, water_mask(img), maskvalues = c(FALSE, NA))
r   <- log_ratio(img)
fit <- fit_bathymetry(extract_ratio(r, pts), max_depth = -16)
depth_map <- predict_bathymetry(fit, r)