satbathy: Satellite-Derived Bathymetry with the Empirical Ratio Transform

Tools to estimate water depth in shallow, low-turbidity coastal waters from multispectral satellite imagery (e.g. Landsat, Sentinel-2 and PlanetScope) using the empirical log-ratio transform of Stumpf, Holderied and Sinclair (2003) doi:10.4319/lo.2003.48.1_part_2.0547 . The package provides water masking, sun-glint correction, spatial filtering, log-ratio computation, linear model fitting and prediction, iterative estimation of the maximum reliable depth, turbidity indexing for image selection, and accuracy assessment.

See also

Author

Maintainer: Jean François Mas jfmas@ciga.unam.mx

Authors:

Jean François Mas jfmas@ciga.unam.mx

Azucena Pérez Vega